Add files via upload
attempted to make a model to predict stats of ceedee lamb for a given week, but there are too many attributes, I am looking to XGBoost to make a more accurate model. The main.py now has adjustments which include headshots for each of the players, their game log when selected, highlighted color for the fantasy points attribute, which indicates how well a player performed in a given week.